Delete all Custom Resources during uninstallation

+++ parallel-install/pkg/deployment/deletion.go
@@ -127,7 +127,10 @@ func (i *Deletion) startKymaUninstallation(prerequisitesEng *engine.Engine, comp
 		return err
 	}
 
-	return i.deleteKymaCrds()
+	if !i.cfg.KeepCRDs {
+		return i.deleteKymaCrds()
+	}
+	return nil
 }
 
 func (i *Deletion) uninstallComponents(ctx context.Context, cancelFunc context.CancelFunc, phase InstallationPhase, eng *engine.Engine, cancelTimeout time.Duration, quitTimeout time.Duration) error {
@@ -239,6 +242,68 @@ func (i *Deletion) deleteKymaNamespaces(namespaces []string) error {
 	finishedCh := make(chan bool)
 	errorCh := make(chan error)
 
+	// All the hacks below should be deleted after this issue is done: https://github.com/kyma-project/kyma/issues/11298
+	//HACK: Delete finalizers of leftover Secret
+	secrets, err := i.kubeClient.CoreV1().Secrets(v1.NamespaceAll).List(context.Background(), metav1.ListOptions{LabelSelector: "serverless.kyma-project.io/config=credentials"})
+	if err != nil && !apierr.IsNotFound(err) {
+		errorCh <- err
+	}
+	if secrets != nil {
+		for _, secret := range secrets.Items {
+			if len(secret.GetFinalizers()) > 0 {
+				secret.SetFinalizers(nil)
+				if _, err := i.kubeClient.CoreV1().Secrets(secret.GetNamespace()).Update(context.Background(), &secret, metav1.UpdateOptions{}); err != nil {
+					errorCh <- err
+				}
+				i.cfg.Log.Infof("Deleted finalizer from \"%s\" Secret", secret.GetName())
+			}
+		}
+	}
+
+	//HACK: Delete finalizers of leftover Custom Resources
+	selector, err := i.prepareKymaCrdLabelSelector()
+	if err != nil {
+		errorCh <- err
+	}
+
+	crds, err := i.apixClient.CustomResourceDefinitions().List(context.Background(), metav1.ListOptions{LabelSelector: selector.String()})
+	if err != nil && !apierr.IsNotFound(err) {
+		errorCh <- err
+	}
+
+	if crds != nil {
+		for _, crd := range crds.Items {
+			customResource := schema.GroupVersionResource{
+				Group:    crd.Spec.Group,
+				Version:  crd.Spec.Version,
+				Resource: crd.Spec.Names.Plural,
+			}
+
+			customResourceList, err := i.dClient.Resource(customResource).Namespace(v1.NamespaceAll).List(context.Background(), metav1.ListOptions{})
+			if err != nil && !apierr.IsNotFound(err) {
+				errorCh <- err
+			}
+			if customResourceList != nil {
+				for _, cr := range customResourceList.Items {
+					if len(cr.GetFinalizers()) > 0 {
+						cr.SetFinalizers(nil)
+						_, err := i.dClient.Resource(customResource).Namespace(cr.GetNamespace()).Update(context.Background(), &cr, metav1.UpdateOptions{})
+						if err != nil {
+							errorCh <- err
+						}
+						i.cfg.Log.Infof("Deleted finalizer for \"%s\" %s", cr.GetName(), cr.GetKind())
+					}
+					if !i.cfg.KeepCRDs {
+						err = i.dClient.Resource(customResource).Namespace(cr.GetNamespace()).Delete(context.Background(), cr.GetName(), metav1.DeleteOptions{})
+						if err != nil && !apierr.IsNotFound(err) {
+							errorCh <- err
+						}
+					}
+				}
+			}
+		}
+	}
+
 	// start deletion in goroutines
 	for _, namespace := range namespaces {
 		err := retry.Do(func() error {
@@ -251,7 +316,7 @@ func (i *Deletion) deleteKymaNamespaces(namespaces []string) error {
 			if len(pods.Items) > 0 {
 				for _, pod := range pods.Items {
 					if pod.Status.Phase == v1.PodRunning {
-						return errors.New(fmt.Sprintf("Namespace %s could not be deleted because of the running Pod: %s. Trying again..", namespace, pod.Name))
+						return errors.New(fmt.Sprintf("\"%s\" Namespace could not be deleted because of the running \"%s\" Pod. Trying again..", namespace, pod.Name))
 					}
 				}
 			}
@@ -259,64 +324,13 @@ func (i *Deletion) deleteKymaNamespaces(namespaces []string) error {
 		}, i.retryOptions...)
 
 		if err != nil {
-			i.cfg.Log.Infof("Namespace %s could not be deleted because of running Pod(s)", namespace)
+			i.cfg.Log.Infof("\"%s\" Namespace could not be deleted because of running Pod(s)", namespace)
 			wg.Done()
 			continue
 		}
 
 		go func(ns string) {
 			defer wg.Done()
-			// All the hacks below should be deleted after this issue is done: https://github.com/kyma-project/kyma/issues/11298
-			//HACK: Delete finalizers of leftover Secret
-			secrets, err := i.kubeClient.CoreV1().Secrets(ns).List(context.Background(), metav1.ListOptions{LabelSelector: "serverless.kyma-project.io/config=credentials"})
-			if err != nil && !apierr.IsNotFound(err) {
-				errorCh <- err
-			}
-			if secrets != nil {
-				for _, secret := range secrets.Items {
-					if len(secret.GetFinalizers()) > 0 {
-						secret.SetFinalizers(nil)
-						if _, err := i.kubeClient.CoreV1().Secrets(ns).Update(context.Background(), &secret, metav1.UpdateOptions{}); err != nil {
-							errorCh <- err
-						}
-						i.cfg.Log.Infof("Deleted finalizer from Secret: %s", secret.Name)
-					}
-				}
-			}
-
-			//HACK: Delete finalizers of leftover Custom Resources
-			crds, err := i.apixClient.CustomResourceDefinitions().List(context.Background(), metav1.ListOptions{LabelSelector: "origin=kyma"})
-			if err != nil && !apierr.IsNotFound(err) {
-				errorCh <- err
-			}
-
-			if crds != nil {
-				for _, crd := range crds.Items {
-					customResource := schema.GroupVersionResource{
-						Group:    crd.Spec.Group,
-						Version:  crd.Spec.Version,
-						Resource: crd.Spec.Names.Plural,
-					}
-
-					customResourceList, err := i.dClient.Resource(customResource).Namespace(ns).List(context.Background(), metav1.ListOptions{})
-					if err != nil && !apierr.IsNotFound(err) {
-						errorCh <- err
-					}
-					if customResourceList != nil {
-						for _, cr := range customResourceList.Items {
-							if len(cr.GetFinalizers()) > 0 {
-								cr.SetFinalizers(nil)
-								_, err := i.dClient.Resource(customResource).Namespace(ns).Update(context.Background(), &cr, metav1.UpdateOptions{})
-								if err != nil {
-									errorCh <- err
-								}
-								i.cfg.Log.Infof("Deleted finalizer from %s: %s", cr.GetKind(), cr.GetName())
-							}
-						}
-					}
-				}
-			}
-
 			err = retry.Do(func() error {
 				//remove namespace
 				if err := i.kubeClient.CoreV1().Namespaces().Delete(context.Background(), ns, metav1.DeleteOptions{}); err != nil && !apierr.IsNotFound(err) {
@@ -329,13 +343,13 @@ func (i *Deletion) deleteKymaNamespaces(namespaces []string) error {
 				} else if apierr.IsNotFound(err) {
 					return nil
 				}
-				i.cfg.Log.Infof("Namespace '%s' still exists with Phase: '%s'", nsT.Name, nsT.Status.Phase)
-				return fmt.Errorf("Namespace '%s' still exists with Phase: '%s'", nsT.Name, nsT.Status.Phase)
+				i.cfg.Log.Infof("\"%s\" Namespace still exists in \"%s\" Phase", nsT.Name, nsT.Status.Phase)
+				return fmt.Errorf("\"%s\" Namespace still exists in \"%s\" Phase", nsT.Name, nsT.Status.Phase)
 			}, i.retryOptions...)
 			if err != nil {
 				errorCh <- err
 			}
-			i.cfg.Log.Infof("Namespace '%s' is removed", ns)
+			i.cfg.Log.Infof("\"%s\" Namespace is removed", ns)
 		}(namespace)
 	}
